script: add script to generate reproducible yearly feeds
This script, for a given commit, generates reproducible yearly feeds.
The output can be uncompressed (but huge), or compressed with either
gzip or xz. The compressed gzip archives are not reproducible (that's a
limitation of the python gzip module, where it is not possible to force
all compression parameters). The xz compressed archives were found to be
reproducible when run on the same machine, but as far as we could see,
there is no such guarantee in the python lzma module, so we do not
advertise they are.

Here are a few timings and sizes (on the 2024-02-25 dataset):

                    Size        Time    Reproducible
    uncompressed    1.5G     46.613s        √
    gzip -6         128M     57.361s        ×
    gzip -9         122M     73.772s        ×
    xz -6           58M     214.396s        √
    xz -8           54M     231.523s        √
    xz -9|extreme   47M     510.218s        √

So, as a trade-off between speed and size, we keep the compression level
6 for both gzip and xz; xz is made the default compression, as it's the
format use in the releases so far.

Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>